disabler() has side effect of disabling executable build
I am not sure if this is a meson bug, or not.  When executable(dependencies)
contains an object that is filled by disabled() the executable is not built.
This makes using disabler() in meson_options.txt conditional else case
impossible, because what I want is to build the executable without that
library.  So for now as a go-around add an unlikely-ever-to-exist dependency
to avoid breakage caused by none-existing variable.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>